+++ src/ts/Ads/Ads.ts
@@ -82,7 +82,7 @@ import { Analytics } from 'Analytics/Analytics';
 import { PrivacySDK } from 'Privacy/PrivacySDK';
 import { PrivacyParser } from 'Privacy/Parsers/PrivacyParser';
 import { Promises } from 'Core/Utilities/Promises';
-import { LoadExperiment, LoadRefreshV4, MediationCacheModeAllowedTest, AuctionXHR } from 'Core/Models/ABGroup';
+import { LoadExperiment, LoadRefreshV4, MediationCacheModeAllowedTest, AuctionXHR, AuctionV6Test, LoadV5 } from 'Core/Models/ABGroup';
 import { PerPlacementLoadManagerV4 } from 'Ads/Managers/PerPlacementLoadManagerV4';
 import { PrivacyMetrics } from 'Privacy/PrivacyMetrics';
 import { PrivacySDKUnit } from 'Ads/AdUnits/PrivacySDKUnit';
@@ -95,7 +95,10 @@ import { createMeasurementsInstance } from 'Core/Utilities/TimeMeasurements';
 import { XHRequest } from 'Core/Utilities/XHRequest';
 import { NofillImmediatelyManager } from 'Ads/Managers/NofillImmediatelyManager';
 import { LegacyCampaignManager } from 'Ads/Managers/LegacyCampaignManager';
+import { AutomatedExperimentManager } from 'Ads/Managers/AutomatedExperimentManager';
 import { CampaignAssetInfo } from 'Ads/Utilities/CampaignAssetInfo';
+import { AdRequestManager } from 'Ads/Managers/AdRequestManager';
+import { PerPlacementLoadManagerV5 } from 'Ads/Managers/PerPlacementLoadManagerV5';
 
 export class Ads implements IAds {
 
@@ -128,13 +131,16 @@ export class Ads implements IAds {
     private _loadApiEnabled: boolean = false;
     private _webViewEnabledLoad: boolean = false;
     private _nofillImmediately: boolean = false;
+    private _forceLoadV5: boolean = false;
     private _mediationName: string;
     private _core: ICore;
+    private _automatedExperimentManager: AutomatedExperimentManager;
 
     public BannerModule: BannerModule;
     public AR: AR;
     public Analytics: Analytics;
     public Store: IStore;
+    public AdRequestManager: AdRequestManager;
 
     constructor(config: unknown, core: ICore) {
         this.PrivacySDK = PrivacyParser.parse(<IRawAdsConfiguration>config, core.ClientInfo, core.DeviceInfo);
@@ -180,29 +186,30 @@ export class Ads implements IAds {
         this.MissedImpressionManager = new MissedImpressionManager(this._core.Api);
         this.ContentTypeHandlerManager = new ContentTypeHandlerManager();
         this.ThirdPartyEventManagerFactory = new ThirdPartyEventManagerFactory(this._core.Api, this._core.RequestManager);
+        this._automatedExperimentManager = new AutomatedExperimentManager();
     }
 
     public initialize(): Promise<void> {
-        this.setupLoadApiEnabled();
-        const measurements = createMeasurementsInstance(InitializationMetric.WebviewInitializationPhases, {
-            'wel': `${this._webViewEnabledLoad}`
-        });
-
-        let promise = Promise.resolve();
-
-        this._nofillImmediately = CustomFeatures.isNofillImmediatelyGame(this._core.ClientInfo.getGameId()) && !!(performance && performance.now);
-        if (this._nofillImmediately) {
-            promise = this.setupMediationTrackingManager().then(() => {
-                this.NofillImmediatelyManager = new NofillImmediatelyManager(this.Api.LoadApi, this.Api.Listener, this.Api.Placement, this.Config.getPlacementIds());
-                return this._core.Api.Sdk.initComplete();
-            });
-        }
-
-        return promise.then(() => {
+        const measurements = createMeasurementsInstance(InitializationMetric.WebviewInitializationPhases);
+        return Promise.resolve().then(() => {
             SdkStats.setInitTimestamp();
             GameSessionCounters.init();
             Diagnostics.setAbGroup(this._core.Config.getAbGroup());
             return this.setupTestEnvironment();
+        }).then(() => {
+            this.setupLoadApiEnabled();
+            measurements.overrideTag('wel', `${this._webViewEnabledLoad}`);
+
+            let promise = Promise.resolve();
+
+            this._nofillImmediately = CustomFeatures.isNofillImmediatelyGame(this._core.ClientInfo.getGameId()) && !!(performance && performance.now);
+            if (this._nofillImmediately) {
+                promise = this.setupMediationTrackingManager().then(() => {
+                    this.NofillImmediatelyManager = new NofillImmediatelyManager(this.Api.LoadApi, this.Api.Listener, this.Api.Placement, this.Config.getPlacementIds());
+                    return this._core.Api.Sdk.initComplete();
+                });
+            }
+            return promise;
         }).then(() => {
             measurements.measure('setup_environment');
             return this.Analytics.initialize();
@@ -255,8 +262,8 @@ export class Ads implements IAds {
             const parserModules: AbstractParserModule[] = [
                 new AdMob(this._core, this),
                 new Display(this._core, this),
-                new MRAID(this.AR.Api, this._core, this),
-                new Performance(this.AR.Api, this._core, this),
+                new MRAID(this.AR.Api, this._core, this._automatedExperimentManager, this),
+                new Performance(this.AR.Api, this._core, this._automatedExperimentManager, this),
                 new VAST(this._core, this),
                 new VPAID(this._core, this),
                 new XPromo(this._core, this)
@@ -271,9 +278,10 @@ export class Ads implements IAds {
                 }
             });
 
-            RequestManager.configureAuctionProtocol(this._core.Config.getTestMode());
+            RequestManager.configureAuctionProtocol(this._core.Config.getTestMode(), AuctionV6Test.isValid(this._core.Config.getAbGroup()));
 
             this.configureCampaignManager();
+            this.configureAutomatedExperimentManager();
             this.configureRefreshManager();
             SdkStats.initialize(this._core.Api, this._core.RequestManager, this._core.Config, this.Config, this.SessionManager, this.CampaignManager, this._core.MetaDataManager, this._core.ClientInfo, this._core.CacheManager);
 
@@ -331,10 +339,30 @@ export class Ads implements IAds {
     }
 
     private configureCampaignManager() {
+        if (this._loadApiEnabled && this._webViewEnabledLoad) {
+            if (this.isLoadV5Enabled()) {
+                this.AdRequestManager = new AdRequestManager(this._core.NativeBridge.getPlatform(), this._core, this._core.Config, this.Config, this.AssetManager, this.SessionManager, this.AdMobSignalFactory, this._core.RequestManager, this._core.ClientInfo, this._core.DeviceInfo, this._core.MetaDataManager, this._core.CacheBookkeeping, this.ContentTypeHandlerManager, this.PrivacySDK, this.PrivacyManager);
+                this.CampaignManager = this.AdRequestManager;
+                return;
+            }
+        }
+
         this.CampaignManager = new LegacyCampaignManager(this._core.NativeBridge.getPlatform(), this._core, this._core.Config, this.Config, this.AssetManager, this.SessionManager, this.AdMobSignalFactory, this._core.RequestManager, this._core.ClientInfo, this._core.DeviceInfo, this._core.MetaDataManager, this._core.CacheBookkeeping, this.ContentTypeHandlerManager, this.PrivacySDK, this.PrivacyManager, this.MediationLoadTrackingManager);
     }
 
+    private configureAutomatedExperimentManager() {
+        if (AutomatedExperimentManager.isAutomationAvailable(this.Config, this._core.Config)) {
+            this._automatedExperimentManager.initialize(this._core, this.CampaignManager.onCampaign);
+        }
+    }
+
     private configureRefreshManager(): void {
+        // AdRequestManager will be set only if Load V5 is enabled.
+        if (this.AdRequestManager) {
+            this.RefreshManager = new PerPlacementLoadManagerV5(this.Api, this.Config, this._core.Config, this.AdRequestManager, this._core.ClientInfo, this._core.FocusManager);
+            return;
+        }
+
         if (this._loadApiEnabled && this._webViewEnabledLoad) {
             const abGroup = this._core.Config.getAbGroup();
             const isZyngaDealGame = CustomFeatures.isZyngaDealGame(this._core.ClientInfo.getGameId());
@@ -381,6 +409,10 @@ export class Ads implements IAds {
                         } else {
                             SDKMetrics.reportMetricEvent(MiscellaneousMetric.XHRNotAvailable);
                         }
+                    } else if (this.isLoadV5Enabled() && this._webViewEnabledLoad) {
+                        experimentType = MediationExperimentType.LoadV5;
+                    } else if (AuctionV6Test.isValid(this._core.Config.getAbGroup())) {
+                        experimentType = MediationExperimentType.AuctionV6;
                     }
 
                     this._mediationName = mediation.getName()!;
@@ -400,7 +432,7 @@ export class Ads implements IAds {
         }
 
         if (CustomFeatures.sampleAtGivenPercent(1)) {
-            Diagnostics.trigger('consent_show', {adsConfig: JSON.stringify(this.Config.getDTO())});
+            Diagnostics.trigger('consent_show', { adsConfig: JSON.stringify(this.Config.getDTO()) });
         }
 
         if (this._core.Config.isCoppaCompliant()) {
@@ -489,6 +521,14 @@ export class Ads implements IAds {
             return;
         }
 
+        // AdRequestManager is only if Load V5 is enabled.
+        // isInvalidationPending is valid only when Load V5 is in use.
+        if (this.AdRequestManager && placement.isInvalidationPending()) {
+            this.showError(true, placementId, 'Invalidation pending for a placement');
+            SDKMetrics.reportMetricEvent(ErrorMetric.PlacementInvalidationPending);
+            return;
+        }
+
         SdkStats.sendShowEvent(placementId);
 
         if (campaign.isExpired()) {
@@ -559,7 +599,7 @@ export class Ads implements IAds {
             AbstractPrivacy.createBuildInformation(this._core.NativeBridge.getPlatform(), this._core.ClientInfo, this._core.DeviceInfo, campaign, this._core.Config);
             this._currentAdUnit = this.getAdUnitFactory(campaign).create(campaign, placement, orientation, playerMetadataServerId || '', options);
             this.RefreshManager.setCurrentAdUnit(this._currentAdUnit, placement);
-            this._currentAdUnit.onClose.subscribe(() =>  {
+            this._currentAdUnit.onClose.subscribe(() => {
                 this.onAdUnitClose();
                 SDKMetrics.sendBatchedEvents();
             });
@@ -718,6 +758,10 @@ export class Ads implements IAds {
         if (TestEnvironment.get('debugJsConsole')) {
             MRAIDView.setDebugJsConsole(TestEnvironment.get('debugJsConsole'));
         }
+
+        if (TestEnvironment.get('forceLoadV5')) {
+            this._forceLoadV5 = true;
+        }
     }
 
     private logChinaMetrics() {
@@ -764,9 +808,17 @@ export class Ads implements IAds {
         const isFanateeExtermaxGameForLoad = CustomFeatures.isFanateeExtermaxGameForLoad(this._core.ClientInfo.getGameId());
         const isOriginalLoad = LoadExperiment.isValid(this._core.Config.getAbGroup()) && CustomFeatures.isWhiteListedForLoadApi(this._core.ClientInfo.getGameId());
         const isLoadV4 = LoadRefreshV4.isValid(this._core.Config.getAbGroup()) && CustomFeatures.isWhiteListedForLoadApi(this._core.ClientInfo.getGameId());
+        const loadV5 = this.isLoadV5Enabled();
 
-        if (isOriginalLoad || isLoadV4 || isZyngaDealGame || isMopubTestGame || isCheetahTestGame || isFanateeExtermaxGameForLoad) {
+        if (isOriginalLoad || isLoadV4 || isZyngaDealGame || isMopubTestGame || isCheetahTestGame || isFanateeExtermaxGameForLoad || loadV5) {
             this._webViewEnabledLoad = true;
         }
+    }
+
+    private isLoadV5Enabled(): boolean {
+        const loadV5Test = LoadV5.isValid(this._core.Config.getAbGroup());
+        const loadV5Game = CustomFeatures.isLoadV5Game(this._core.ClientInfo.getGameId());
+
+        return (loadV5Test && loadV5Game) || this._forceLoadV5;
     }
 }