+++ src/ts/Ads/Parsers/AuctionResponseParser.ts
@@ -25,15 +25,21 @@ export class AuctionResponseParser {
     private static constructTrackingUrls(trackingTemplates: string[], tracking: IPlacementTrackingV6): ICampaignTrackingUrls {
         const trackingUrls: ICampaignTrackingUrls = {};
 
-        const globalParams = tracking.params;
-        const events = tracking.events;
+        const globalParams = tracking.params || {};
+        const events = tracking.events || {};
 
-        Object.keys(events).forEach(((eventKey) => {
-            const eventTracking = events[eventKey];
-            const params = {
-                ...(eventTracking.params || {}),
+        Object.keys(events).forEach(((eventKey: string) => {
+            const eventTracking = events[eventKey] || {};
+            const tempParams: { [x: string]: string } = {
+                ...eventTracking.params,
                 ...globalParams
             };
+            const params: { [x: string]: string } =  {};
+            Object.keys(tempParams).forEach(key =>  {
+                const templateKey = `{{${key}}}`;
+                params[templateKey] = tempParams[key];
+            });
+
             eventTracking.urlIndices.forEach((index) => {
                 if (index >= 0 && index < trackingTemplates.length) {
                     const tempTrackingUrls: string[] = trackingUrls[eventKey] || [];