+++ src/ts/MRAID/AdUnits/MRAIDAdUnitParametersFactory.ts
@@ -37,15 +37,12 @@ export class MRAIDAdUnitParametersFactory extends AbstractAdUnitParametersFactor
     private _webPlayerContainer: WebPlayerContainer;
     private _automatedExperimentManager: AutomatedExperimentManager;
 
-    constructor(ar: IARApi, core: ICore, ads: IAds) {
+    constructor(ar: IARApi, core: ICore, ads: IAds, automatedExperimentManager: AutomatedExperimentManager) {
         super(core, ads);
-        this._automatedExperimentManager = new AutomatedExperimentManager(core);
+        this._automatedExperimentManager = automatedExperimentManager;
 
         if (ads.Config.getHasArPlacement()) {
-            this._automatedExperimentManager.initialize(ArAutomatedExperimentsList).catch(() => {
-                SDKMetrics.reportMetricEvent(AUIMetric.AutomatedExperimentManagerInitializationError);
-            });
-            this._automatedExperimentManager.beginExperiment();
+            this._automatedExperimentManager.registerExperiments(ArAutomatedExperimentsList);
         }
 
         this._ar = ar;
@@ -67,8 +64,7 @@ export class MRAIDAdUnitParametersFactory extends AbstractAdUnitParametersFactor
             if ((resourceUrl && resourceUrl.getOriginalUrl().match(/playables\/production\/unity/)) || MRAIDAdUnitParametersFactory._forcedExtendedMRAID) {
                 mraid = new ExtendedMRAID(baseParams.platform, baseParams.core, baseParams.deviceInfo, baseParams.placement, baseParams.campaign, baseParams.privacy, showGDPRBanner, baseParams.coreConfig.getAbGroup(), baseParams.gameSessionId, baseParams.adsConfig.getHidePrivacy());
             } else if (ARUtil.isARCreative(baseParams.campaign) || MRAIDAdUnitParametersFactory._forcedARMRAID) {
-                const decision = arAvailableButtonDecision(this._automatedExperimentManager);
-
+                const decision = arAvailableButtonDecision(this._automatedExperimentManager, baseParams.campaign);
                 mraid = new ARMRAID(baseParams.platform, baseParams.core, this._ar, baseParams.deviceInfo, baseParams.placement, baseParams.campaign, baseParams.deviceInfo.getLanguage(), baseParams.privacy, showGDPRBanner, baseParams.coreConfig.getAbGroup(), baseParams.gameSessionId, baseParams.adsConfig.getHidePrivacy(), this._automatedExperimentManager, decision);
             } else if (baseParams.campaign.isCustomCloseEnabled()) {
                 mraid = new CustomCloseMRAID(baseParams.platform, baseParams.core, baseParams.deviceInfo, baseParams.placement, baseParams.campaign, baseParams.privacy, showGDPRBanner, baseParams.coreConfig.getAbGroup(), baseParams.gameSessionId, baseParams.adsConfig.getHidePrivacy());