+++ test/Unit/Ads/UserPrivacyManagerTest.ts
@@ -76,7 +76,9 @@ describe('UserPrivacyManagerTest', () => {
         userPrivacy.getPermissions.returns({
            ads: false,
            gameExp: false,
-           external: false
+           external: false,
+           agreedOverAgeLimit: false,
+           agreementMethod: ''
         });
         privacySDK.getUserPrivacy.returns(userPrivacy);
 
@@ -926,7 +928,8 @@ describe('UserPrivacyManagerTest', () => {
                    ads: false,
                    external: false,
                    gameExp: false,
-                   agreedOverAgeLimit: false
+                   agreedOverAgeLimit: false,
+                   agreementMethod: ''
                });
             });