+++ CHANGELOG.md
@@ -1,14 +1,50 @@
-# Pending
+# Wed Apr  1 15:21:28 UTC 2020
 
+* Privacy SDK metrics callback [#14041](https://github.com/Applifier/unity-ads-webview/pull/14041)
+
+# Tue Mar 31 20:51:59 UTC 2020
+
+* [Auction V6] Enable AB Test on Group 18 [#14080](https://github.com/Applifier/unity-ads-webview/pull/14080)
+* [Auction V6] Fix Templatizing Parameters [#14135](https://github.com/Applifier/unity-ads-webview/pull/14135)
+* Disable fire_deploy_json.sh During LKG Updates [#14136](https://github.com/Applifier/unity-ads-webview/pull/14136)
+
+# Tue Mar 31 18:29:12 UTC 2020
+
+* Post WebView size to GitHub as commit status [#14072](https://github.com/Applifier/unity-ads-webview/pull/14072)
+* Replace OMIDPartner macro in VastURITag unwrapping for IAS [#14075](https://github.com/Applifier/unity-ads-webview/pull/14075)
+* Force LKGs to Update Locally for Revert Script [#14134](https://github.com/Applifier/unity-ads-webview/pull/14134)
+* [Auction V6] Add Test Mode Endpoint Support [#14079](https://github.com/Applifier/unity-ads-webview/pull/14079)
+* [Auction V6] Add Mediation Experiment Type [#14106](https://github.com/Applifier/unity-ads-webview/pull/14106)
+* [Load V5] Metadata flag to force Load V5 to be enabled [#13920](https://github.com/Applifier/unity-ads-webview/pull/13920)
+* [Load V5] Move loadCampaign logic from AdRequestManager to PerPlacementLoadManagerV5 [#14073](https://github.com/Applifier/unity-ads-webview/pull/14073)
+* [Load V5] Invalidation pending state for a placement [#14077](https://github.com/Applifier/unity-ads-webview/pull/14077)
+* [Load V5] Add comment about AdRequestManager [#14078](https://github.com/Applifier/unity-ads-webview/pull/14078)
+* [Auction V6] Add Support for Empty Tracking [#14076](https://github.com/Applifier/unity-ads-webview/pull/14076)
+
+# Reverted to LKGs (Mon Mar 30 21:57:00 UTC 2020) - Tue Mar 31 15:44:19 UTC 2020
+
+* Reverted [Auction V6] Add Support for Empty Tracking [#14076](https://github.com/Applifier/unity-ads-webview/pull/14076)
+* Reverted Add Initial Optimized Campaign Support [#12639](https://github.com/Applifier/unity-ads-webview/pull/12639)
+
+# Mon Mar 30 21:57:00 UTC 2020
+
+* [Auction V6] Add Support for Empty Tracking [#14076](https://github.com/Applifier/unity-ads-webview/pull/14076)
+* Add Initial Optimized Campaign Support [#12639](https://github.com/Applifier/unity-ads-webview/pull/12639)
+
+# Fri Mar 27 23:32:15 UTC 2020
+
+* [Auction V6] Add Disabled ABTest [#14070](https://github.com/Applifier/unity-ads-webview/pull/14070)
 * [Auction V6] Add Auction Response Parser [#14033](https://github.com/Applifier/unity-ads-webview/pull/14033)
 * [Auction V6] Use Auction Parser in LegacyCampaignManager [#14034](https://github.com/Applifier/unity-ads-webview/pull/14034)
 * [Auction V6] Add Metric Tracking For Failure Cases [#14035](https://github.com/Applifier/unity-ads-webview/pull/14035)
+* [Load V5] Add Disabled ABTest for Selected Games [#13890](https://github.com/Applifier/unity-ads-webview/pull/13890)
 * [Load V5] Add Support for PerPlacementLoadManagerV5 [#13889](https://github.com/Applifier/unity-ads-webview/pull/13889)
 * [Load V5] Use reload campaign if ongoing load was canceled [#14068](https://github.com/Applifier/unity-ads-webview/pull/14068)
 * [Load V5] Set Retry to 0 for Requests [#14038](https://github.com/Applifier/unity-ads-webview/pull/14038)
 * [Load V5] Mocks for PerPlacementLoadManagerV5 tests [#14039](https://github.com/Applifier/unity-ads-webview/pull/14039)
 * [Load V5] Increment Ad Ordinal [#14036](https://github.com/Applifier/unity-ads-webview/pull/14036)
 * [Load V5] Parameters Test for Load V5 [#13990](https://github.com/Applifier/unity-ads-webview/pull/13990)
+* [Load V5] Configure test environment as fast as possible [#14069](https://github.com/Applifier/unity-ads-webview/pull/14069)
 * Open Measurement - Add Timeout on Admob start metric [#14040](https://github.com/Applifier/unity-ads-webview/pull/14040)
 * Update AndroidDeviceInfo DTO [#14001](https://github.com/Applifier/unity-ads-webview/pull/14001)